refactor: move ComPtr to main library
